requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
frozenlist==1.4.0
gTTS==2.3.2
idna==3.4
install==1.3.5
multidict==6.0.4
openai==0.27.8
pefile==2023.2.7
PyAudio==0.2.13
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.6
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
requests==2.31.0
SpeechRecognition==3.10.0
tqdm==4.65.0
urllib3==2.0.4
yarl==1.9.2
